requirements.txt
Django>=3.2.5
djangorestframework>=3.14.0
pandas>=1.5.1
numpy>=1.23.4
requests>=2.28.1
beautifulsoup4>=4.9.3
openpyxl>=3.0.9
mysql>=0.0.3
mysql-connector-python>=8.0.31
python-dotenv>=0.21.0
cache>=1.0.3
diskcache>=5.4.0
mysqlclient>=2.1.0
fontawesomefree>=6.2.0
pytest>=5.4.0
pymongo>=4.3.2
pika>=1.3.1
celery>=5.2.7
